fix(ui)remove deleted apiConfigs from stateManager
earlier: if an api is deleted using ui, it is no more tracked
by the executor, but statemanger maintains an entry for it,
always showing its status as RED

now: statemanager takes a param `CleanDur`. If executor doesn't
send any event for an api A during last `CleanDur` duration, it is
deleted from statemanager also.
    point to note here is that this cleanup operation kicks in only
when Get-State-API is called from UI.